fix: 更新好友群組和好友申請的資料驗證邏輯
- 在 friendApplication.js 和 friendGroup.js 中將註解從「獲取資料」更新為「驗證資料」，以提高代碼的可讀性和一致性。
- 增強資料驗證邏輯，確保在創建、更新和刪除好友群組時，只有擁有相應權限的用戶才能進行操作，提升系統的安全性。
- 在 get.js 中新增獲取好友群組的功能，簡化資料存取流程，確保資料的正確性和完整性。

+++ socket/friendGroup.js
@@ -14,9 +14,16 @@ const {
 const friendGroupHandler = {
   createFriendGroup: async (io, socket, data) => {
     try {
-      const { group: newGroup, userId } = data;
-      console.log(newGroup, userId);
-      if (!newGroup || !userId) {
+      // data = {
+      //   userId: string,
+      //   group: {
+      //     ...
+      //   },
+      // }
+
+      // Validate data
+      const { group: _newFriendGroup, userId } = data;
+      if (!_newFriendGroup || !userId) {
         throw new StandardizedError(
           '無效的資料',
           'ValidationError',
@@ -25,30 +32,57 @@ const friendGroupHandler = {
           401,
         );
       }
+      const newFriendGroup = await Func.validate.friendGroup(_newFriendGroup);
 
+      // Validate socket
       const operatorId = await Func.validate.socket(socket);
 
+      // Get data
+      const operator = await Get.user(operatorId);
+      const user = await Get.user(userId);
+      const userSocket = Object.values(io.sockets.sockets).find(
+        (s) => s.userId === user.id,
+      );
+
+      // Validate operation
+      if (operator.id !== user.id) {
+        throw new StandardizedError(
+          '您沒有權限新增非自己的好友群組',
+          'ValidationError',
+          'CREATEFRIENDGROUP',
+          'PERMISSION_DENIED',
+          403,
+        );
+      }
+
+      // Create friend group
       const friendGroupId = uuidv4();
       await Set.friendGroup(friendGroupId, {
-        ...newGroup,
-        userId,
-        createdAt: new Date(),
+        ...newFriendGroup,
+        userId: user.id,
+        createdAt: Date.now(),
       });
 
-      io.to(socket.id).emit('userUpdate', {
-        friendGroups: await Get.userFriendGroups(operatorId),
+      // Emit updated data (to the user)
+      io.to(userSocket.id).emit('userUpdate', {
+        friendGroups: await Get.userFriendGroups(user.id),
       });
 
       new Logger('FriendGroup').success(
-        `FriendGroup(${friendGroupId}) created`,
+        `FriendGroup(${friendGroupId}) of User(${user.id}) created by User(${operator.id})`,
       );
     } catch (error) {
       if (!(error instanceof StandardizedError)) {
         error = new StandardizedError(
           `新增好友群組時發生無法預期的錯誤: ${error.message}`,
+          'ServerError',
+          'CREATEFRIENDGROUP',
+          'EXCEPTION_ERROR',
+          500,
         );
       }
 
+      // Emit error data (to the operator)
       io.to(socket.id).emit('error', error);
 
       new Logger('FriendGroup').error(
@@ -58,12 +92,17 @@ const friendGroupHandler = {
   },
 
   updateFriendGroup: async (io, socket, data) => {
-    // Get database
-    const friendGroups = (await db.get('friendGroups')) || {};
-
     try {
-      const { friendGroupId, group } = data;
-      if (!friendGroupId || !group) {
+      // data = {
+      //   friendGroupId: string,
+      //   group: {
+      //     ...
+      //   },
+      // }
+
+      // Validate data
+      const { friendGroupId, group: _editedFriendGroup } = data;
+      if (!friendGroupId || !_editedFriendGroup) {
         throw new StandardizedError(
           '無效的資料',
           'ValidationError',
@@ -72,30 +111,55 @@ const friendGroupHandler = {
           401,
         );
       }
-
-      const operatorId = await Func.validate.socket(socket);
-      const friendGroup = await Func.validate.friendGroup(
-        friendGroups[friendGroupId],
+      const editedFriendGroup = await Func.validate.friendGroup(
+        _editedFriendGroup,
       );
 
-      await Set.friendGroup(friendGroup.id, {
-        ...group,
-      });
+      // Validate socket
+      const operatorId = await Func.validate.socket(socket);
 
-      io.to(socket.id).emit('userUpdate', {
-        friendGroups: await Get.userFriendGroups(operatorId),
+      // Get data
+      const operator = await Get.user(operatorId);
+      const friendGroup = await Get.friendGroup(friendGroupId);
+      const user = await Get.user(friendGroup.userId);
+      const userSocket = Object.values(io.sockets.sockets).find(
+        (s) => s.userId === friendGroup.userId,
+      );
+
+      // Validate operation
+      if (operator.id !== friendGroup.userId) {
+        throw new StandardizedError(
+          '您沒有權限修改非自己的好友群組',
+          'ValidationError',
+          'UPDATEFRIENDGROUP',
+          'PERMISSION_DENIED',
+          403,
+        );
+      }
+
+      // Update friend group
+      await Set.friendGroup(friendGroup.id, editedFriendGroup);
+
+      // Emit updated data (to the user)
+      io.to(userSocket.id).emit('userUpdate', {
+        friendGroups: await Get.userFriendGroups(user.id),
       });
 
       new Logger('FriendGroup').success(
-        `FriendGroup(${friendGroup.id}) updated`,
+        `FriendGroup(${friendGroup.id}) of User(${user.id}) updated by User(${operator.id})`,
       );
     } catch (error) {
       if (!(error instanceof StandardizedError)) {
         error = new StandardizedError(
           `更新好友群組時發生無法預期的錯誤: ${error.message}`,
+          'ServerError',
+          'UPDATEFRIENDGROUP',
+          'EXCEPTION_ERROR',
+          500,
         );
       }
 
+      // Emit error data (to the operator)
       io.to(socket.id).emit('error', error);
 
       new Logger('FriendGroup').error(
@@ -105,10 +169,12 @@ const friendGroupHandler = {
   },
 
   deleteFriendGroup: async (io, socket, data) => {
-    // Get database
-    const friendGroups = (await db.get('friendGroups')) || {};
-
     try {
+      // data = {
+      //   friendGroupId: string,
+      // }
+
+      // Validate data
       const { friendGroupId } = data;
       if (!friendGroupId) {
         throw new StandardizedError(
@@ -120,27 +186,51 @@ const friendGroupHandler = {
         );
       }
 
+      // Validate socket
       const operatorId = await Func.validate.socket(socket);
-      const friendGroup = await Func.validate.friendGroup(
-        friendGroups[friendGroupId],
+
+      // Get data
+      const operator = await Get.user(operatorId);
+      const friendGroup = await Get.friendGroup(friendGroupId);
+      const user = await Get.user(friendGroup.userId);
+      const userSocket = Object.values(io.sockets.sockets).find(
+        (s) => s.userId === friendGroup.userId,
       );
 
+      // Validate operation
+      if (operator.id !== friendGroup.userId) {
+        throw new StandardizedError(
+          '您沒有權限刪除非自己的好友群組',
+          'ValidationError',
+          'DELETEFRIENDGROUP',
+          'PERMISSION_DENIED',
+          403,
+        );
+      }
+
+      // Delete friend group
       await db.delete(`friendGroups.${friendGroup.id}`);
 
-      io.to(socket.id).emit('userUpdate', {
-        friendGroups: await Get.userFriendGroups(operatorId),
+      // Emit updated data (to the user)
+      io.to(userSocket.id).emit('userUpdate', {
+        friendGroups: await Get.userFriendGroups(user.id),
       });
 
       new Logger('FriendGroup').success(
-        `FriendGroup(${friendGroup.id}) deleted`,
+        `FriendGroup(${friendGroup.id}) of User(${user.id}) deleted by User(${operator.id})`,
       );
     } catch (error) {
       if (!(error instanceof StandardizedError)) {
         error = new StandardizedError(
           `刪除好友群組時發生無法預期的錯誤: ${error.message}`,
+          'ServerError',
+          'DELETEFRIENDGROUP',
+          'EXCEPTION_ERROR',
+          500,
         );
       }
 
+      // Emit error data (to the operator)
       io.to(socket.id).emit('error', error);
 
       new Logger('FriendGroup').error(
